refactor: using new throwError function in auth middleware and fixing imports in users controller and service
