+++ middleware/authentication.js
@@ -1,6 +1,7 @@
 const { validateToken } = require('../functions/jsonwebtoken');
 const { FORBIDDEN: FORBIDDEN_CODE } = require('../constants/httpStatus');
 const { FORBIDDEN: FORBIDDEN_MESSAGE } = require('../constants/message');
+const throwError = require("../functions/throw-error");
 
 module.exports = (req, res, next) => {
   // Sends the request to validate the token
@@ -10,7 +11,7 @@ module.exports = (req, res, next) => {
   //   - the token didn't exist in the request
   //   - the token wasn't valid
   //   - the token expired
-  if (payload === null) return res.status(FORBIDDEN_CODE).send(FORBIDDEN_MESSAGE);
+  if (payload === null) return throwError(FORBIDDEN_CODE, FORBIDDEN_MESSAGE)
 
   // Adds the payload of the token to the request. This will let access
   // it from another place without revalidating the token
